Fix the database fields

+++ app/database/budget_plan.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Sequence, ForeignKey, CheckConstraint
+from sqlalchemy import Column, Integer, Sequence, ForeignKey, CheckConstraint, Numeric
 
 
 def create_budget_table(db):
@@ -6,7 +6,7 @@ def create_budget_table(db):
         __tablename__ = "budget_plan"
         id = Column(Integer, Sequence("seq_income_plan_id", start=1), primary_key=True)
         user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-        amount = Column(Integer, nullable=False)
+        amount = Column(Numeric(10, 2), nullable=False)
         category_id = Column(Integer, ForeignKey("category.id"), nullable=False)
         year = Column(Integer, nullable=False)
         month = Column(Integer, nullable=False)
