+++ app/database/category.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, Sequence, ForeignKey, Date
+from sqlalchemy import Column, Integer, String, Boolean, Sequence, ForeignKey, Date, Numeric
 
 
 def create_category_table(db):
@@ -27,7 +27,7 @@ def create_goals_table(db):
         __tablename__ = "goals"
         id = Column(Integer, Sequence("seq_goals", start=1), primary_key=True)
         name = Column(String, nullable=False)
-        target_amount = Column(Integer)
+        target_amount = Column(Numeric(10, 2))
         deadline = Column(Date)
         user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
